+++ composeApp/src/commonMain/kotlin/com/katya/app/AppVersion.kt
@@ -1,6 +1,6 @@
 package com.katya.app
 
 object AppVersion {
-    const val APP_VERSION: String = "3.1.1"
-    const val VERSION_CODE: Int = 144
+    const val APP_VERSION: String = "3.1.3-git"
+    const val VERSION_CODE: Int = 145
 }